refactor(commit): 作業ディレクトリ生成処理をcreate-workdir.tsへ分離
- commit-prepare.tsから作業ディレクトリ生成関連の関数・定数を抽出
- create-workdir.tsにtimestamp関数やcreateWorkdirPath等を移動
- コードの責務を整理し可読性・保守性を向上
